Examined 00_on_actions file
I'm not sure if I got absolutely everything that was wrong, but there were quite a few events that simply do not exist in AtE but which were still referenced in this file. (Note that some event files haven't been copied over to AtE because they don't have to - they lack references to religion, culture, or region. so they don't have to be copied over.)